pyproject.toml
[project]
name = "dimity-jones-solutions"
version = "0.1.0"
description = "Solutions for Dimity Jones in Puzzle Castle https://obnakwa.itch.io/dimityjones"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.1",
    "python-constraint>=1.4.0",
]
